Enable SAVEQUERIES when the debug bar is enabled
Luckily, WP doesn't define this constant, so we can define it ourselves
(if it hasn't been defined previously, of course).

This is a little late, and will probably miss some queries, but the
majority are going to be happening here.

At some point, we need to move this out of here into an earlier loading
spot - this is more a stopgap solution