Always set Jetpack to staging mode outside the VIP environment.
This is a first pass at a fix for #127 and doesn't cover every case.
Namely, it doesn't cover staging sites hosted on VIP Go. The biggest
risk right now is developers downloading a production backup from
VaultPress which includes the Jetpack connection information, and
restoring that to their local environment. So long as they also use our
mu-plugins, this would avoid breaking the Jetpack connection in that
case.