Switch filename uniqueness checks to 4.5's new filter for this purpose
Currently, we use a variable hook, which has led to issue with client and plugin code that changes the action. WP 4.5 introduced a filter in `wp_unique_filename()` that we can use instead, ensuring that regardless of the `action` passed, the file service is checked for uniqueness.

There's a lot of repetition in this changeset, which will be addressed.

Also, the new filter doesn't allow error responses or check file type, so a new callback will be introduced to handle that, within the Core function that checks filetypes already.