Cache Purging: More sensible purge limits
Trying to purge too many URLs at once slows down post saving and will eventually cause post save actions to fail. We can be a bit more conservative with the number of pages that we purge. Luckily, this method of manually sending purge requests to each cache node is temporary and the new method should be much more robust.