Filter WP API urls to always use HTTPS
Always use HTTPS on WP API urls, regardless of other site settings.

This applies to the production as well as local development
environments, to keep theme as close as possible.

If this must be disabled for local development, the filter can be
removed, but HTTPS is enforced at the web server level in production.