a8c-files.php: use option_upload_url_path
This commit changes the `pre_option_upload_url_path` hook to `option_upload_url_path` in order to allow other plugins to modify the option. Now we are also not changin the value in case it is not empty.

Per feedback in https://github.com/Automattic/vip-go-mu-plugins/pull/376#discussion_r91154368