a8c-files.php: change the upload_url_path even if it is empty
Not applying our filter to an option in case it is not empty does actually not bring anything good and may make the code and it's behaviour less predictable.

In case a plugin wants to filter the option, it should do it on higher priority than 10 at which we operate.