Prevent Jetpack version ping-pong on sandboxes
When a sandbox has an old version of stacks, this can lead to some
serious issues because Jetpack tries to auto-correct the version on each
load. The mismatch between the sandbox environment and live environment
leads to a constant fight to override the option.

This just kills a request when an attempt to do a Jetpack version update
is triggered from a sandbox.