Adding caching support for is_multi_author
`is_multi_author()` normally caches itself if the site has multiple authors, but if not, it tries to cache a `false` value which always makes the lookup happen.  This can cause a very slow query on sites where many tens of thousands of posts are attributed to the same author.

Can be used by adding this filter to the theme: `add_filter( 'pre_transient_is_multi_author', 'wpcom_vip_is_multi_author' );`

This helper function will cache the results for one hour as a string so even failures will be cached, since `is_multi_author()` casts its output as a boolean.