Update to reduce slow media queries
Ticket about slow queries on Laughing Squid: https://wordpressvip.zendesk.com/agent/tickets/65236

The slowness is the result of a select-distinct query when looking for media. Select-distinct is already expensive so it should be cached. But if it doesn't get any results, it still has to do a bunch of complicated processing, returns nothing, and nothing isn't cached, so the query stays slow. 

This filter caches the nothing, as it were. It is active on Wordpress.com and tests okay on LaughingSquid.com, a VIP site. 

Here's context for the bug:  https://keepingtheirblogsgoing.wordpress.com/2016/04/27/speeding-up-wp-admin/
Here's the same code that was pushed on wp.com on 19 April: https://wpcom.trac.automattic.com/changeset/154832