Add cache-busting functions
The previous commit cached an array of months and years that have media attachments. 

We need to make sure the transient is updated to reflect which months and years do, in fact, have media attachments in them. wpcom_vip_media_clear_caches and wpcom_vip_bust_media_months_cache will do that by clearing the cache when a new video or audio attachment is uploaded. The first function deletes a cached check for whether the post is a video or audio, the second deletes and refreshes the cached months array.