Adding vip-cache.php
This file will hopefully contain helper scripts for cache related commands.

Right now, I've started it with a command to purge a site's Varnish cache.  This is useful for automating site cloning, so a support user doesn't have to log in to every newly cloned site and manually purge the cache.  It's pretty barebones (it's got one job) but I'm open to suggestions of other additions if needed.