Fix is_ssl() for wp-cli
Based on the scheme for the site's siteurl.

This is because we need to make sure the correct URLs are
returned by the site_url() function, which relies on is_ssl() and
defaults to false in our currently environment (i.e. sites with an
https siteurl will end up with an http `site_url()`)