Move all New Relic Logic to a separate file
Add 2 functions that are added to plugins_loaded to detect when WP_CRON is true and when WP_CLI is true and to mark them as separate transactions in New Relic. We will also not take these transactions into account when calculating the apdex score since those are not end user facing.