Adding image-delta WP CLI command
As there might be cases in which some images are imported to environment different from production, we need a way for moving the image delta to production upon content sync.

The newly introduced `wp vip migration image-delta` command makes it easy to import the image delta right after the SQL has been brought from preprod or dev to prod.

Using the `--uploaded-after-id` or `--uploaded-after-date` params allows you to import only images, which made it to preprod/dev after the initial import and thus import only the delta.

The ID or a date of the latest post from intial SQL import has to be determined manually.