Lower the wp_calculate_image_srcset_meta filter priority
A lot of other plugins and custom implementations are hooking the `wp_calculate_image_srcset_meta` filter on default priority (10), either implicitly, or explicitly.

Since this implementation is going to be platform's default, we should hook prior those custom ones, in order to make absolutely sure the custom implementations override this solution.