Replace the CALCULATE_IMAGE_SRCSET_META constant for a filter
We want to let clients to control the functionality, if needed, via filter. In order to do so from a theme, we need to init the whole thing on init action.

This commit replaces the CALCULATE_IMAGE_SRCSET_META constant for a wpcom_enable_srcset_and_sizes filter (the same as what was used on WPCOM).

This commit also contains a default `add_filter( 'wpcom_enable_srcset_and_sizes', '__return_false' );` line for the period of time in between deployment of the code and official rollaout for all clients.

Clients will be able to enabled the functionality prior the hard launch either in all envs, or in selected ones.