Jetpack: Enforce some modules.
The `jetpack_active_modules` filter allows developers to enable/disable some modules in code. We might need to ensure some modules are always enabled/disabled and so we hook in to that filter ourselves.

At the moment, this just forces Vaultpress to be enabled. It won't have any effect until https://github.com/Automattic/jetpack/pull/8465 is merged and released.