Run WP CLI commands via WP Cron
This is not ready for client use. If you are interested in learning, please contact VIP Support.

The code in this PR is a mechanism for offloading WP CLI commands to WP Cron which allows those command to be run on CLI containers, instead of sandboxes, which takes the commands closer to the production servers and thus should provide better performance for the commands.