Moving the sprintf to act on `$paging_endpoint` only
In case the slug contains some percentage signs, for instance due to encoding non-ascii chars, the `sprintf` call fails.

Instead of proposed str_replace escaping ( #823 ), this approach does no mess up with the URL, but only calls sprintf with `$paging_endpoint` as a param - that variable is supposed to contain  single `%d` conversion specifier and thus the `sprintf` should not fail due to the confusion over missing params.

Fixes #823