WP_Filesystem compatability layer
We currently don't have a good way to interact with the file system on VIP Go. This commit allows us to use some of the functions provided by WP_Filesystem.
The first layer is a compatability layer that detects if we're trying to do an operation on the /tmp folder or in the uploads folder. Everything else will trigger a E_USER_ERROR. It also handled doing things like copying from tmp to uploads and vice versa.
For the tmp folder we use WP_Filesystem_Direct() and for the VIP Go we use WP_Filesystem_VIP_GO() which in turn calls the VIP GO files API.