Improving handling of non-image attachments
The non-image attachments, eg.: PDFs, are taking advantage of fallback_sites, filterable via `fallback_intermediate_image_sizes`, which generate the thumbnails.

Those were not properly filtered out on VIP Go and the lack of the prevention of those was borking the srcset implementation tests.

Besides adding the filter, the tests for PDF attachment revealed some improper checks around the code. This commit addresses PHP notices and tightens the checks.