Bubble up errors
Since the errors come from the WP_Filesystem_VIP_Uploads class we need to bubble them up to the WP_Filesystem_VIP class. The way the errors work is that everytime a new error occurs it's added to the internal $this->errors variable. Since the variable will contain all of the previous errors, including any new ones, we will just overwrite the WP_Filesystem_VIP with the one from the transport that was used.