Only enforce restricted usernames on VIP
Currently, the restrcited usernames functionality enforces usernames on every single environment that the mu-plugins library is used on, including local. However, quite a lot of local environments use the admin username by default, which makes it impossible to login with a default configuration. Chassis, in particular, sets up WordPress in this way and it is a recommended local environment for developing on VIP. This simple change only enforces the restricted usernames when on a WPCOM_VIP environment instead of doing so indiscriminately.