Fix hash_equals() php warning (#977)
`hash_equals()` expects both values passed into it to be strings.
However `get_proxy_verification_key()` may return an integer
when `WPCOM_VIP_PROXY_VERIFICATION` is undefined,
triggering a PHP warning. Casting `rand()` to a  string prevents this.

phpunit tests updated to reflect this casting on the return.
`assertNotEmpty()` used to verify that the string of numbers are properly
returned when `WPCOM_VIP_PROXY_VERIFICATION` undefined or blank.