Liveblog: Purge Varnish for each new liveblog entry.
Given that Varnish caches posts for 30 minutes, a liveblog post can contain important meta data that is as much as 30 minutes out of date. This is because the post includes data about the last updated timestamp and this is included in the cached output.

This data can therefore differ from the actual entries, which are loaded asynchronously.