Add support for basic auth to internal API endpoints
Use case: We need to use the vip/v1/sites endpoint to get information
about multisites in the migration CLI. Rather than hardcoding a shared
secret in the CLI, we can create a user on the site and use that
username and password to authenticate for the API endpoints.

This uses the basic auth plugin (https://github.com/WP-API/Basic-Auth)
to populte the current user based on HTTP basic auth. We can then use
that current user to decide whether to authenticate the current request
for our internal endpoints.

Note: We're also checking `current_user_can( 'vip_support' )` here
because VIP Support users only have `read`, `subscriber`, and
`vip_support` in this context. We may just want to fix that instead.

Testing:

1. Load the PR on your sandbox
2. Create support user on a test site
3. curl -u 'username:password' https://<domain>/wp-json/vip/v1/sites