Don't fire off deletes to the files service for every image size
Use a static variable to record whether we've deleted a file before.

Of course there could be cases when more than one file is deleted in a single request. The assumption here is that all variations of the same image would be deleted in a row, thereby only allowing one through with the next image resetting the static variable on it's first variation.

It should be noted that it is not an issue to let extra deletes through, this is just to reduce unnecessary calls and speed up the transaction.