Change approach on WP_Filesystem_VIP_Uploads->sanitize_uploads_path()
The problem with the first approach was that it restricted users to _only_ being able to write to their subsite's folder within `/wp-content/uploads` which isn't ideal - especially considering a plugin that needs to clone sites/media.

I've essentially stripped $upload_basedir back down to `/tmp/uploads` from `/tmp/uploads/sites/{site_id}` for the search/replace that happens later in the code. This allows users to pass in any path they want _without_ losing any parts of the path.

Does anyone see any pitfalls in this approach? Or is there a smarter way to do this?