Require the GUTENBERG_USE_PLUGIN constant
For WordPress 5.0+, any environments that want to use the Gutenberg plugin need to define the `GUTENBERG_USE_PLUGIN` constant. This facilitates the upgrade path where core disables the Gutenberg plugin as part of the upgrade. Without the constant in place, the upgrade routine would fail to deactivate the plugin and continue using it.

If the constant is defined, any attempts to load the plugin via `wpcom_vip_load_plugin` will skip loading the plugin and just fallback to the core block editor.