Update WP_Filesystem_VIP->is_uploads_path() logic for multisite
On multisite, since you can access the `/wp-content/uploads` directory for any site and `wp_get_upload_dir()['basedir']` is now corrected to include the `/sites/{site_id}` portion of the URL, the existing `WP_Filesystem_VIP->is_uploads_path()` logic is flawed.

If you passed in the path of a file from a different site in the network, `WP_Filesystem_VIP->is_uploads_path()` would return `false` and would result in `WP_Filesystem_Direct` being chosen for the transport method causing the file action to fail. Since the only real criteria needed to determine if the path is in the local uploads directory or not is whether the file path contains `/tmp/uploads`, I've updated it to just check for that value instead.

If there are common scenarios where the uploads path on VIP Go is _not_ `/tmp/uploads` then this may need to be tweaked. I'm using `get_temp_dir()` and appending 'uploads' to it, so if the uploads directory will _always_ be `{whatever the /tmp directory is}/uploads`, we should still be fine here.