Reset core roles on migration cleanup if any are missing
Rather than just checking for `administrator`, this checks if any core roles are missing. We have to be careful here because if any new core roles are added, we'd have to update this list. It may be better just to `populate_roles()` every time we run migration cleanup. That's also not perfect -- if a site has removed any core roles, this will re-add them.