Fix partner-provision.sh on CLI containers
After migrations, we run `wp vip migration cleanup` on a CLI container.
This runs `wp jetpack-start connect` internally, which calls the
partner-provision script.

Apparently the PATH variable is not set in the PHP exec(), which breaks
the partner-provision script. Here we're setting the full path to
executables that are used in the partner-provision script.

Alternatively, we could fix the PATH environment variable.

Related: https://github.com/Automattic/vip-go-api/pull/1702