Time gate forced 2FA to 1 month from now
Before this date, we just output an admin notice to users that will need to enable 2FA. The logic to display the date got a little complex, but
ensuring the notice is in their local timezone is worth it.

After the date specified, the admin notice switches to say 2FA is required and only the most basic functions are available.

This is a date that's a month from now in the middle-ish of the day in the middle of the week (so plenty of support will be around in case there are questions).