Fixes attachments with completely missing meta
If an attachment exists, but has no attachment meta, we should still add the filesize.  This is already set up to happen in `Meta_Updater::update_attachment_filesize()` but the check in `Meta_Updater::get_attachments()` does not allow the attachment ID to be fixed if it does not already have attachment metadata.