Don't enforce 2FA by default in local dev
Adds a `wpcom_vip_is_two_factor_local_testing` filter. 2FA enforcement will be disabled by default in local environments, but if you need to test the 2FA functionality, you can easily enable it in local environments with this filter.

We generally try to make the local environment as close as possible to production, but this is an instance where it would just get in the way most of the time.