Abstract Jetpack authorize logic
Abstracts the logic that detects XML-RPC and REST Jetpack Authorization
requests to wpcom_vip_is_jetpack_authorize_request(). The goal is to
make wpcom_vip_should_force_two_factor slightly easier to grok. You
don't necessarily have to care about the details of how we're detecting
the authorization requests, just that we're not enforcing 2fa when we do.