Debug Bar: Play nice with other enable filters
The existing filter callback to force-enable the Debug Bar for Automatticians was a little too black-and-white. If you are an Automattician, it was enabled, and if you aren't, then it got disabled.

This doesn't work well with filter callbacks that clients might have, that may enable or disable it based on role, capability, or environment.

If they ran at an earlier priority (say, 50), then their conditions get ignored, as the vip-mu-go-plugins code doesn't take any of it into consideration.

If they ran at a later priority, then their conditions would also need to take `is_automattician()` into account, or otherwise Debug Bar gets disabled for VIP Support users; that isn't something that the client code should care about.

The solution is to only force-enable Debug Bar for Automatticians, but NOT force-disable if they are not - instead, let whatever value that might have been set by an earlier callback be maintained.

This doesn't stop a later client callback from trouncing all over the VIP logic, but with this callback being more accommodating, they should no longer need to run their callback at a later priority; I've bumped the priority of our callback to hopefully ensure that we run after any existing client callbacks.