WP-CLI: When `url` not set, force the main site
We routinely run into cases where a mismatch in the wp-config and the
multisite config leads to failures on WP_CLI (which can make debugging,
fixing the issue, and other work challenging).

This forces loads the "main site" based on the IDs specified in config
when the url flag is not set. This way we avoid "site not found" errors
when a site is in a mismatched state and allows WP_CLI to continue
working.