Cache Manager: Purge posts page
If the site is using a different page to list posts, i.e. `page_for_posts`, we should purge that posts page URL.

We limit it to when the post type is a `post`, to avoid unnecessarily clearing an index of `post` posts, when a custom post type was what was updated.

Fixes #241.