PHPCS: Remove file exclusions
While the code in these files would fail `WordPress-Core` standard, that is not something we're checking for, so there is no reason to exclude them. They do not have any violations according to PHPCompatibilityWP and WordPress-VIP-Go standards.

If WPCS was added later on, then there is a better way to exclude them for WPCS, rather than excluding them from _all_ PHPCS checks (including PHPCompatibilityWP and WordPress-VIP-Go).

See #1032.