Remove invalid tests
Now that we are defaulting the constant in code, we can't use set it dynamically at runtime in the code.

The change will become default in JP 8.2, so the tests will be obsolete soon, and getting them to work correctly within the confines of the current constant-based config approach is more effort than needed.