Add documents that failed to index to queue
Most indexing happens via bulk indexing in ElasticPress. So hooking on
that filter captures new posts and what have you. There's not a lot of
granularity, so all posts that were a part of any failure have to be
added to the queue to be reindexed just to be safe.