Refactor the WooCommerce admin access check.
@mjangda suggested a simpler approach to checking whether or not WooCommerce should let this user access `wp-admin`.

Here, we have an array of capabilities that WooCommerce checks for in allowing access to `wp-admin` and a static variable to track whether or not we've already granted this user access.

First, if we haven't yet granted access and the current `$cap` is one of the WC caps to check, we get the current user's capabilities, check if `$cap` is set and if it's `true`. If so, this user should be able to access `wp-admin`. We update our static variable so we don't check any other capabilities, and add the `woocommerce_prevent_admin_access` filter to `__return_false`.