Filter `woocommerce_prevent_admin_access` to not block valid users from wp-admin (#1576)
* Filter `woocommerce_prevent_admin_access` to not block valid users from wp-admin.

To determine if `wp-admin` access should be blocked, WooCommerce is checking the capabilities on the user and not the user's actual role. Since the VIP 2FA plugin alters the user's capabilities so early, when WooCommerce checks them to see if the user should be able to access `wp-admin`, valid users who should have `wp-admin` access are being blocked and redirected to the front end of the site.

I don't think it's unreasonable for WooCommerce to assume that a user's capabilities _should_ match the capabilites of their role. We can use the `woocommerce_prevent_admin_access` filter to introduce a strict check to see that a user's _role_ has the capabilities that WooCommerce checks for and determine whether or not the user should have access to `wp-admin`.

This commit adds that filter, returns early if the VIP 2FA plugin is not enabled OR if WooCommerce has already determined that the user should have `wp-admin` access. It also introduces a helper function that takes an array of capabilites, looks through all roles for that capability, and returns an array of role slugs for _any_ role that contains _any_ of the passed capabilities. If the current user's role matches any role in the returned array, we grant that user access to `wp-admin`.

* Whitespace fixes

* Return false instead of $prevent_access

As per suggestion, returning `false` for clarity.

* Refactor the WooCommerce admin access check.

@mjangda suggested a simpler approach to checking whether or not WooCommerce should let this user access `wp-admin`.

Here, we have an array of capabilities that WooCommerce checks for in allowing access to `wp-admin` and a static variable to track whether or not we've already granted this user access.

First, if we haven't yet granted access and the current `$cap` is one of the WC caps to check, we get the current user's capabilities, check if `$cap` is set and if it's `true`. If so, this user should be able to access `wp-admin`. We update our static variable so we don't check any other capabilities, and add the `woocommerce_prevent_admin_access` filter to `__return_false`.