Send batched purges to the `purge-batch` endpoint
Group purge requests in batches of `CACHE_PURGE_BATCH_SIZE` and send the
batched purge requests to the `purge-batch` endpoint.

Batched requests are used only if the `PURGE_BATCH_SERVER_URL` constant
is defined in `wp-config.php`, otherwise the code falls back to the old
approach of sending individual requests to the `purge` endpoint.

When batched requests are supported, the number of URLs that can be
purged is increased to 4000, while it remains 100 in the fallback case.