is_proxied_automattician: Check constant exists
The `is_proxied_request()` checks whether the `A8C_PROXIED_REQUEST` constant is defined, before checking it is `=== true`.

The `is_proxied_automattician()` only checked whether the constant was truthy. It did not check if the constant was defined.

It makes sense for `is_proxied_automattician()` (the purpose of which is to check for both a proxied request AND that the currently logged-in user appears to be an Automattician) to use `is_proxied_request()` internally.

Note, that technically this slightly tightens the behaviour, from allowing a truthy defintion (like `'a string'` or `1`) to requiring it to be a boolean `true`.

The constant may be defined in secret file as true, but [falls back](https://github.com/Automattic/vip-go-mu-plugins/blob/4511c412617032b71e820f7adc997c9358c35905/000-vip-init.php#L59-L64) to `false` if not. Considering VIP control the secrets file, and `is_proxied_request()` already checks `=== true`, then it seems reasonable to carry this behaviour through to `is_proxied_automattician()` as well.