Run headless Chrome without a sandbox
The circleci container does not seem to support running a sandboxed
headless Chrome, and puppeteer tests throw the following error:
`No usable sandbox! If you want to live dangerously and need an
immediate workaround, you can try using --no-sandbox.`

Since the container runs on circleci and only performs local tests the
fix is to disable running Chrome in a sandbox.