Define environment variable within jobs
When using anchors as aliases, YAML does not include keys from the
anchor that are already defined in the entity. This means that the
`WP_CORE_DIR` and WP_TESTS_DIR` env variables would be undefined when
running the job.

A quick fix is to define the `environment` variables at the job level,
instead of including them in the anchor.