QM: Re-add previously ignored output directory
In #1683, the Query Monitor plugin was changed from being a git
submodule to a standard set of files. The version of the plugin was not
changed.

However, the `output` directory was missing from #1683, due to a rule in
the `.gitignore` file.

This PR adds an exception to that rule for the query-monitor/output
directory, and adds the missing directory itself.

Fixes #1684.