Cron Control Next: Merge jobs back into a single cache entry
The idea here is to limit the number of cache requests. Since we're only
updating 1 row at a time now, a stale local cache doesn't matter, so
we no longer have to bypass the local cache. This means we can make 1
request for all cron control jobs at once instead of individual
requests.

https://github.com/Automattic/Cron-Control/pull/178/commits/f0189fb9bbadb95fd69bafde2ec00179baa83880