Disable do_pings on sites that use `pre_schedule_event`
The new (pending) version of Cron Control utilizes the `pre_*` hooks to hook into cron, allowing for a much cleaner way to move cron to its own table. When that plugin is active, we hook on `pre_schedule_event` which short-circuits the schedule_event function before we disable do_pings.

Here we're disabling on both pre_schedule_event as well as schedule_event for backwards compatibility.