Add APC Cache Interceptor
The APC Cache Interceptor defines hot cache keys to be intercepted and persisted in APCu.

The idea here is to create some short-term persistent caches to reduce load on memcached.

Our first two examples here are

* `userlogins:wpcomvip` which is requested on every page and just returns a user ID that never changes
* `sites:<id>` which returns some basic information about a subsite and very rarely changes

The caveat is that changes are not propagated to other PHP instances, so cached data is potentially
stale as soon as it is cached. It is important that we don't use stale data to overwrite the database. This
is why it's notable that the values above don't (or very rarely) change and why we're only caching the site
data for 10 seconds.

Depends on https://github.com/Automattic/wp-memcached/pull/66