Don't offload subsite information in network admin
Offloading the subsite information in the network admin has the side effect of showing stale data after clicking save on a site. (i.e. from /wp-admin/network/site-info.php?id=<id>). The form still saves properly and if you refresh the page, you see the correct thing. It's kind of suboptimal though.

To counteract that, we're disabling apcu for these keys when you're in the network admin. That way, you still get the data directly from memcached in the screen when you could be updating this data, but in all other cases, you get the apcu offloaded data.