Bump vip-support to https://github.com/Automattic/vip-support/commit/1c47e4c9ae71c966c6c0c275e6f010e1f6fcba4d
Don't add roles on init

add_role can lead to front-end writes which are not ideal. It can lead to undesirable circumstances with other roles being removed if there are intermittent database issues.

We still trigger our upgrade routine on admin_init so new installs will still get the roles (and there's still a chance for the latter issue to trigger but the likelihood is significantly reduced).

Also includes a convenience WP-CLI command to force run the upgrade routine if needed (wp vipsupport reset-roles)