Add additional data to MemcachePool errors.
While we have a maximum size of (compressed) data per individual cache key, right now we have no way of determining what keys might be causing these issues at the application layer.

I'm hoping we can use our custom error handler to extract additional information from the backtrace and add it to the message. I'm also moving this to `E_WARNING` because, in my opinion, this is more in line with the severity of an error like this. Having objects uncached can cause site stability issues.