Removes full-width Non-Prod Bar CSS
The Non-Prod Bar (Sandbox Bar) is not full width, but contains a "hidden" rectangle that extends the width of the site, meaning you can't click anything behind it. This is an annoyance and does not cause any actual functionality issues.

This fix removes the 100% width, thus allowing clicks on any part of the screen. The Non-Prod Bar still displays properly and still works properly when clicked on its visible portion.