Remove WP-Content Writable Test Added in WordPress 6.3
WordPress 6.3 adds the ability for failed plugin and theme update to be automatically rolled back.
The feature requires a temp directory in `wp-content/` to function.

As part of this feature a new Health Check is added to ensure that the temp directory can be created.
On WPVIP updates are managed via GitHub and `wp-content` is read-only so this check will always fail.

This PR removes the check from the Health Check page.

See: https://core.trac.wordpress.org/ticket/51857