Set WP_TEMP_DIR for Contact Form 7 uploads
The plugin ignores the value of `WPCF7_UPLOADS_TMP_DIR` if it is not
within a set of paths allowed by the plugin. The plugin allows the path
defined in `WP_TEMP_DIR`, so this fix uses the plugin's `wpcf7_upload_dir`
filter to set `WP_TEMP_DIR` to `/tmp/` so that the plugin will allow the
path that we defined earlier in `WPCF7_UPLOADS_TMP_DIR`.