Add base telemetry library (take 2) (#5869)
This PR adds a new generic Telemetry library consisting of base classes for Automattic's Tracks system integration. This is based heavily on the existing Telemetry package under the vip-parsely directory but is meant to be more generic and can be used by other plugins, and extended with other telemetry providers. There are no singletons, and most classes have fixed responsibilities.

---------

Co-authored-by: Hanif Norman <hanifnorman@gmail.com>
Co-authored-by: Hanif Norman <1227524+hanifn@users.noreply.github.com>
Co-authored-by: Alec Geatches <alec.geatches@automattic.com>